requirements.txt
amqp==5.1.0
asgiref==3.5.0
async-generator==1.10
async-timeout==4.0.2
attrs==21.4.0
Babel==2.9.1
beautifulsoup4==4.10.0
billiard==3.6.4.0
branca==0.4.2
celery==5.2.6
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.11
click==8.0.4
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.3.1
cryptography==36.0.1
dateutils==0.6.12
decorator==5.1.1
defusedxml==0.7.1
Deprecated==1.2.13
dj-database-url==0.5.0
Django==3.2.13
django-allauth==0.47.0
django-bootstrap5==21.3
django-celery-beat==2.2.1
django-celery-results==2.3.0
django-ckeditor==6.2.0
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-filter==21.1
django-jet-reboot==1.3.1
django-js-asset==1.2.2
django-localflavor==3.1
django-on-heroku==1.1.2
django-paypal==2.0
django-phonenumber-field==6.0.0
django-phonenumbers==1.0.1
django-rest-auth==0.9.5
django-rest-framework-braces==0.3.4
django-svg-templatetag==1.0.1
django-timezone-field==4.2.3
djangorestframework==3.13.1
djangorestframework-simplejwt==5.0.0
feedparser==6.0.8
Flask==2.0.3
flower==1.0.0
folium==0.12.1.post1
future==0.18.2
geocoder==1.38.1
greenlet==1.1.2
gunicorn==20.1.0
h11==0.13.0
haversine==2.5.1
humanize==4.0.0
idna==3.3
imutils==0.5.4
itsdangerous==2.1.1
Jinja2==3.0.3
kombu==5.2.4
localflavor==1.9
MarkupSafe==2.1.1
numpy==1.22.2
oauthlib==3.2.0
opencv-python==4.5.5.62
outcome==1.1.0
packaging==21.3
phonenumbers==8.12.42
Pillow==9.0.1
prometheus-client==0.14.1
prompt-toolkit==3.0.29
psycopg2-binary==2.9.3
pycparser==2.21
pygraphviz==1.9
PyJWT==2.3.0
pyOpenSSL==22.0.0
pyparsing==3.0.8
pytesseract==0.3.8
python-crontab==2.6.0
python-dateutil==2.8.2
python-stdnum==1.17
python3-openid==3.2.0
pytz==2021.3
ratelim==0.1.6
redis==4.2.2
requests==2.27.1
requests-oauthlib==1.3.1
selenium==4.1.0
sgmllib3k==1.0.0
six==1.16.0
sniffio==1.2.0
sortedcontainers==2.4.0
soupsieve==2.3.1
SQLAlchemy==1.4.32
sqlparse==0.4.2
tornado==6.1
trio==0.19.0
trio-websocket==0.9.2
urllib3==1.26.8
vine==5.0.0
wcwidth==0.2.5
webdriver-manager==3.5.4
Werkzeug==2.0.3
whitenoise==5.3.0
wrapt==1.14.0
wsproto==1.0.0
WTForms==3.0.1
